backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
openai
pytest
pytest-asyncio
httpx
langchain
langchain-openai
langchain-community
sqlmodel
faiss-cpu
PyPDF2
python-docx
python-multipart
deepl
googletrans==4.0.0-rc1
spacy